reorganize connection classes
