+++ agent/indy_catalyst_agent/messaging/connections/handlers/connection_request_handler.py
@@ -1,10 +1,11 @@
 from ...base_handler import BaseHandler, BaseResponder, RequestContext
 from ..messages.connection_request import ConnectionRequest
-from ....connection import ConnectionManager
+from ..manager import ConnectionManager
 
 
 class ConnectionRequestHandler(BaseHandler):
     """ """
+
     async def handle(self, context: RequestContext, responder: BaseResponder):
         self._logger.debug(f"ConnectionRequestHandler called with context {context}")
         assert isinstance(context.message, ConnectionRequest)
