+++ agent/indy_catalyst_agent/messaging/request_context.py
@@ -6,8 +6,8 @@ import copy
 import logging
 
 from .agent_message import AgentMessage
+from .connections.models.connection_target import ConnectionTarget
 from .message_factory import MessageFactory
-from ..models.connection_target import ConnectionTarget
 from ..storage.base import BaseStorage
 from ..wallet.base import BaseWallet
