+++ agent/indy_catalyst_agent/messaging/responder.py
@@ -1,7 +1,7 @@
 from abc import ABC
 
 from .agent_message import AgentMessage
-from ..models.connection_target import ConnectionTarget
+from .connections.models.connection_target import ConnectionTarget
 from ..error import BaseError
